requirements.txt
streamlit==1.28.1
pandas==2.1.1
sqlalchemy==2.0.21
pyodbc==4.0.39
pytz==2023.3
python-dotenv==1.0.0
